// frontend/packages/shared/src/openapi-types.ts
export interface operations {
// ...
    getNotes: {
        parameters: {
            query?: never;
            header?: {
                authorization?: string;
            };
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description Response for status 200 */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        id: string;
                        user_id: string;
                        title: string;
                        body: string;
                        body_iv: string;
                        date: string;
                        archived: boolean;
                        version: number;
                        created_at: string;
                        updated_at: string;
                        file?: {
                            note_id?: string;
                            title: string;
                            path: string;
                        };
                    }[];
                };
            };
        };
    };
    deleteNotes: {
        parameters: {
            query?: never;
            header?: {
                authorization?: string;
            };
            path?: never;
            cookie?: never;
        };
        requestBody: {
            content: {
                "application/json": {
                    ids: string[];
                };
                "application/x-www-form-urlencoded": {
                    ids: string[];
                };
                "multipart/form-data": {
                    ids: string[];
                };
            };
        };
        responses: {
            /** @description Response for status 200 */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        id: string;
                    }[];
                };
            };
        };
    };
    getNotesIds: {
        parameters: {
            query?: never;
            header?: {
                authorization?: string;
            };
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description Response for status 200 */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        id: string;
                        title: string;
                    }[];
                };
            };
        };
    };
// ...
    putNote: {
        parameters: {
            query?: never;
            header?: {
                authorization?: string;
            };
            path?: never;
            cookie?: never;
        };
        requestBody: {
            content: {
                "application/json": {
                    /** Format: uuid */
                    id: string;
                    /** Format: uuid */
                    user_id?: string;
                    title: string;
                    body: string;
                    body_iv: string;
                    date: string;
                    archived?: boolean;
                    version?: number;
                };
                "application/x-www-form-urlencoded": {
                    /** Format: uuid */
                    id: string;
                    /** Format: uuid */
                    user_id?: string;
                    title: string;
                    body: string;
                    body_iv: string;
                    date: string;
                    archived?: boolean;
                    version?: number;
                };
                "multipart/form-data": {
                    /** Format: uuid */
                    id: string;
                    /** Format: uuid */
                    user_id?: string;
                    title: string;
                    body: string;
                    body_iv: string;
                    date: string;
                    archived?: boolean;
                    version?: number;
                };
            };
        };
        responses: {
            /** @description Response for status 200 */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        id: string;
                        user_id: string;
                        title: string;
                        body: string;
                        body_iv: string;
                        date: string;
                        archived: boolean;
                        version: number;
                        created_at: string;
                        updated_at: string;
                        file?: {
                            note_id?: string;
                            title: string;
                            path: string;
                        };
                    };
                };
            };
        };
    };
    putNotesById: {
        parameters: {
            query?: never;
            header?: {
                authorization?: string;
            };
            path: {
                id: string;
            };
            cookie?: never;
        };
        requestBody: {
            content: {
                "application/json": {
                    /** Format: uuid */
                    id?: string;
                    /** Format: uuid */
                    user_id?: string;
                    title?: string;
                    body?: string;
                    body_iv?: string;
                    date?: string;
                    archived?: boolean;
                    version?: number;
                };
                "application/x-www-form-urlencoded": {
                    /** Format: uuid */
                    id?: string;
                    /** Format: uuid */
                    user_id?: string;
                    title?: string;
                    body?: string;
                    body_iv?: string;
                    date?: string;
                    archived?: boolean;
                    version?: number;
                };
                "multipart/form-data": {
                    /** Format: uuid */
                    id?: string;
                    /** Format: uuid */
                    user_id?: string;
                    title?: string;
                    body?: string;
                    body_iv?: string;
                    date?: string;
                    archived?: boolean;
                    version?: number;
                };
            };
        };
        responses: {
            /** @description Response for status 200 */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        id: string;
                        user_id: string;
                        title: string;
                        body: string;
                        body_iv: string;
                        date: string;
                        archived: boolean;
                        version: number;
                        created_at: string;
                        updated_at: string;
                        file?: {
                            note_id?: string;
                            title: string;
                            path: string;
                        };
                    };
                };
            };
            /** @description Response for status 404 */
            404: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    type: unknown;
                };
            };
        };
    };
// ...
    putNotesByIdFile: {
        parameters: {
            query?: never;
            header?: {
                authorization?: string;
            };
            path: {
                id: string;
            };
            cookie?: never;
        };
        requestBody: {
            content: {
                "application/json": {
                    title: string;
                    path: string;
                    /** Format: uuid */
                    note_id?: string;
                };
                "application/x-www-form-urlencoded": {
                    title: string;
                    path: string;
                    /** Format: uuid */
                    note_id?: string;
                };
                "multipart/form-data": {
                    title: string;
                    path: string;
                    /** Format: uuid */
                    note_id?: string;
                };
            };
        };
        responses: {
            /** @description Response for status 200 */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        note_id?: string;
                        title: string;
                        path: string;
                    };
                };
            };
            /** @description Response for status 404 */
            404: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    type: unknown;
                };
            };
        };
    };
}